pyproject.toml
[project]
name = "documents"
version = "0.1.0"
description = ""
authors = [
    {name = "John Doe",email = "johndoe@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
